CHANGED:
def latest_version(self):
    versies = self.enkelvoudiginformatieobject_set.order_by("-versie", "-pk")

TO:

def latest_version(self):
    versies = self.enkelvoudiginformatieobject_set.order_by("-versie", "-pk")

Added "-pk" to ordering because the Factory class in the unit test suite creates two EIO, with identical versions. This did not cause any issues with postgres version 10, but this does yield errors in higher postgres versions like 13 or 14. Ordering on pk is fixes this for the moment..